app/requirements.txt
customtkinter==5.2.1
pymysql==1.1.0
python-dotenv==1.0.1
cryptography==42.0.5
